requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
beautifulsoup4==4.12.3
billiard==4.2.0
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
Django==4.2.16
django-celery-beat==2.7.0
django-timezone-field==7.0
djangorestframework==3.15.2
idna==3.8
kombu==5.4.0
phonenumbers==8.13.44
prompt_toolkit==3.0.47
pycryptodome==3.20.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
redis==5.0.8
requests==2.32.3
six==1.16.0
soupsieve==2.6
sqlparse==0.5.1
sw-django-utils==0.0.54
sw-python-utils==0.0.18
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
